requirements.txt
TinyDB
python-telegram-bot==13.15
flask
requests